+++ src/main/java/com/app/TrabJava.java
@@ -1,0 +1,16 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ */
+
+package com.app;
+
+/**
+ *
+ * @author 202108083852
+ */
+public class TrabJava {
+
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
